Move token calculator to account
